+++ modules/apps/headless/headless-delivery/headless-delivery-client/src/main/java/com/liferay/headless/delivery/client/serdes/v1_0/DocumentFolderSerDes.java
@@ -245,13 +245,13 @@ public class DocumentFolderSerDes {
 
 				if (jsonParserFieldValue != null) {
 					documentFolder.setNumberOfDocumentFolders(
-						Long.valueOf((String)jsonParserFieldValue));
+						Integer.valueOf((String)jsonParserFieldValue));
 				}
 			}
 			else if (Objects.equals(jsonParserFieldName, "numberOfDocuments")) {
 				if (jsonParserFieldValue != null) {
 					documentFolder.setNumberOfDocuments(
-						Long.valueOf((String)jsonParserFieldValue));
+						Integer.valueOf((String)jsonParserFieldValue));
 				}
 			}
 			else if (Objects.equals(jsonParserFieldName, "siteId")) {
