+++ modules/apps/headless/headless-delivery/headless-delivery-client/src/main/java/com/liferay/headless/delivery/client/serdes/v1_0/KnowledgeBaseFolderSerDes.java
@@ -264,7 +264,7 @@ public class KnowledgeBaseFolderSerDes {
 
 				if (jsonParserFieldValue != null) {
 					knowledgeBaseFolder.setNumberOfKnowledgeBaseArticles(
-						Long.valueOf((String)jsonParserFieldValue));
+						Integer.valueOf((String)jsonParserFieldValue));
 				}
 			}
 			else if (Objects.equals(
@@ -272,7 +272,7 @@ public class KnowledgeBaseFolderSerDes {
 
 				if (jsonParserFieldValue != null) {
 					knowledgeBaseFolder.setNumberOfKnowledgeBaseFolders(
-						Long.valueOf((String)jsonParserFieldValue));
+						Integer.valueOf((String)jsonParserFieldValue));
 				}
 			}
 			else if (Objects.equals(
